+++ lib/animationSystem/animation_system_delegate_builder.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:feed/animationSystem/state.dart';
+import 'package:feed/util/global/functions.dart';
 import 'package:feed/util/icon_position.dart';
 import 'package:flutter/material.dart';
 import 'package:fort/fort.dart';
@@ -10,11 +11,13 @@ class AnimationSystemDelegateBuilder extends StatefulWidget {
   final AnimationSystemDelegate delegate;
   final AnimationSystemController controller;
   final bool animateAccordingToPosition;
+  final bool controlHeptic;
   const AnimationSystemDelegateBuilder({ 
     Key? key,
     required this.delegate,
     required this.controller,
-    this.animateAccordingToPosition = false
+    this.animateAccordingToPosition = false,
+    this.controlHeptic = false
   }) : super(key: key);
 
   @override
@@ -29,13 +32,14 @@ class _AnimationSystemDelegateBuilderState extends State<AnimationSystemDelegate
   /// Animates the
   late AnimationController animationController;
 
+  int? oldDirection;
+
   @override
   void initState(){
     super.initState();
     // Initiate state
     tower = AnimationSystemState.tower();
-
-    animationController = AnimationController(vsync: this, duration: widget.delegate.duration);
+    animationController = AnimationController(vsync: this, duration: widget.delegate.duration, lowerBound: 0);
   }
 
   @override
@@ -47,20 +51,92 @@ class _AnimationSystemDelegateBuilderState extends State<AnimationSystemDelegate
   @override
   void dispose(){
     super.dispose();
-    widget.controller.dispose();
   }
+
+/*
+ 
+      ____ _____ _____ _____ _____ ____  ____  
+     / ___| ____|_   _|_   _| ____|  _ \/ ___| 
+    | |  _|  _|   | |   | | |  _| | |_) \___ \ 
+    | |_| | |___  | |   | | | |___|  _ < ___) |
+     \____|_____| |_|   |_| |_____|_| \_\____/ 
+                                               
+ 
+*/
+
+  IconPosition iconPosition(DismissDirection direction){
+    switch (direction) {
+      case DismissDirection.startToEnd:
+        return IconPosition.RIGHT;
+      case DismissDirection.endToStart:
+        return IconPosition.LEFT;
+      case DismissDirection.down:
+        return IconPosition.BOTTOM;
+      case DismissDirection.up:
+        return IconPosition.TOP;
+      default:
+        return IconPosition.RIGHT;
+    }
+  }
+
+  CardPosition cardPosition(DismissDirection direction, double dx){
+    if(direction == DismissDirection.startToEnd){
+      return CardPosition.Right;
+    }
+    else if(direction == DismissDirection.endToStart){
+      return CardPosition.Left;
+    }
+    else if(dx >= 0){
+      return CardPosition.Right;
+    }
+    else{
+      return CardPosition.Left;
+    }
+  }
+
+  double get HORIZONTALSWIPETHRESH{
+    return 92.0;
+  }
+
+  double get BOTTOMSWIPETHRESH{
+    return 157;
+  }
+
+  double get TOPSWIPETHRESH{
+    return 92.0;
+  }
+
+/*
+ 
+     _____                 _   _                 
+    |  ___|   _ _ __   ___| |_(_) ___  _ __  ___ 
+    | |_ | | | | '_ \ / __| __| |/ _ \| '_ \/ __|
+    |  _|| |_| | | | | (__| |_| | (_) | | | \__ \
+    |_|   \__,_|_| |_|\___|\__|_|\___/|_| |_|___/
+                                                 
+ 
+*/
 
   /// Reset the animation state, should be reset after onSwipe has completed
   void reset(){
     tower.dispatch(SetAllAnimationValues(null, null, 0, 0));
-    animationController.animateTo(0, duration: Duration(milliseconds: 0));
+    animationController.reset();
     widget.delegate.onUpdate(0, 0, 0);
   }
 
-  /// When the values of the animation system need to be updated from swipefeed (internal)
-  void _onUpdate(double dx, double dy, Function(DismissDirection)? value){
+  Future<void> _reverse() async {
+    _onUpdate(0, 0, (direction) => 0, true);
+    Future.delayed(Duration(milliseconds: 200));
+    return;
+  }
 
-    if(value == null) return;
+  /// When the values of the animation system need to be updated from swipefeed (internal)
+  void _onUpdate(double dx, double dy, Function(DismissDirection)? value, [bool reverse = false]) async {
+
+    if(value == null || !mounted) return;
+
+    //Direction to int
+    int i = -1;
 
     //Height of the screen
     final height = MediaQuery.of(context).size.height;
@@ -86,135 +162,161 @@ class _AnimationSystemDelegateBuilderState extends State<AnimationSystemDelegate
     if(dx.abs() >= 0 && dy > verticalLength*-1 && dy < verticalLength){
       axisLock = Axis.horizontal;
     }
-    else if(dy.abs() > 0 && dx > horizontalLength*-1 && dx < horizontalLength){
+    else if((dy > 98 || dy < -32) && dx > horizontalLength*-1 && dx < horizontalLength){
       axisLock = Axis.vertical;
     }
 
-    if(dx == 0 && dy == 0){
-      tower.dispatch(SetAllAnimationValues(null, null, dx, dy));
-      return;
+    if(dx == 0 && dy == 0 && reverse){
+      tower.dispatch(SetAllAnimationValues(tower.state.iconPosition, tower.state.cardPosition, dx, dy, reversing: true));
+      animationController.animateTo(0, duration: Duration(milliseconds: 200));
+      await Future.delayed(Duration(milliseconds: 200));
     }
-    if(axisLock != Axis.horizontal && !horizontalAxisOverride){
+    else if(dx == 0 && dy == 0){
+      tower.dispatch(SetAllAnimationValues(null, null, dx, dy));
+    }
+    else if(axisLock != Axis.horizontal && !horizontalAxisOverride){
       if(dy > 0){
+        // show bottom
+        i = 2;
         // Card Position Right Vertical Axis, Below Y axis
         if(dx > 0){
+          tower.dispatch(SetAllAnimationValues(IconPosition.BOTTOM, CardPosition.Right, dx, dy));
           if(widget.delegate.animateAccordingToPosition){
-            animationController.animateTo(value(DismissDirection.down), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.down));
+            animationController.animateTo(value(DismissDirection.down), duration: Duration(milliseconds: 0));
           }
           else{
-            animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.startToEnd));
+            animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
           }
-          tower.dispatch(SetAllAnimationValues(IconPosition.BOTTOM, CardPosition.Right, dx, dy));
-          return;
         }
         // Card Position Left Vertical Axis, Below Y axis
         else{
+          // show top
+          tower.dispatch(SetAllAnimationValues(IconPosition.BOTTOM, CardPosition.Left, dx, dy));
           if(widget.delegate.animateAccordingToPosition){
-            animationController.animateTo(value(DismissDirection.down), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.down));
+            animationController.animateTo(value(DismissDirection.down), duration: Duration(milliseconds: 0));
           }
           else{
-            animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.endToStart));
+            animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
           }
-          tower.dispatch(SetAllAnimationValues(IconPosition.BOTTOM, CardPosition.Left, dx, dy));
-          return;
         }
       }
       else{
+        // show top
+        i = 3;
         // Card Position Right Vertical Axis, above Y axis
         if(dx > 0){
+          tower.dispatch(SetAllAnimationValues(IconPosition.TOP, CardPosition.Right, dx, dy));
           if(widget.delegate.animateAccordingToPosition){
-            animationController.animateTo(value(DismissDirection.up), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.up));
+            animationController.animateTo(value(DismissDirection.up), duration: Duration(milliseconds: 0));
           }
           else{
-            animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.startToEnd));
+            animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
           }
-          tower.dispatch(SetAllAnimationValues(IconPosition.TOP, CardPosition.Right, dx, dy));
-          return;
         }
         // Card Position Left Vertical Axis, above Y axis
         else{
+          tower.dispatch(SetAllAnimationValues(IconPosition.TOP, CardPosition.Left, dx, dy));
           if(widget.delegate.animateAccordingToPosition){
-            animationController.animateTo(value(DismissDirection.up), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.up));
+            animationController.animateTo(value(DismissDirection.up), duration: Duration(milliseconds: 0));
           }
           else{
-            animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
             widget.delegate.onUpdate(dx, dy, value(DismissDirection.endToStart));
+            animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
           }
-          tower.dispatch(SetAllAnimationValues(IconPosition.TOP, CardPosition.Left, dx, dy));
-          return;
         }
       }
     }
     else if(axisLock != Axis.vertical){
       // Card Position Horizontal Axis Right
       if(dx > 0){
-        animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
-        widget.delegate.onUpdate(dx, dy, value(DismissDirection.startToEnd));
+        // show right
+        i = 0;
         tower.dispatch(SetAllAnimationValues(IconPosition.RIGHT, CardPosition.Right, dx, dy));
-        return;
+        widget.delegate.onUpdate(dx, dy, value(DismissDirection.startToEnd));
+        animationController.animateTo(value(DismissDirection.startToEnd), duration: Duration(milliseconds: 0));
       }
       // Card Position Horizontal Axis Left
       else{
-        animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
+        // show left
+        i = 1;
         widget.delegate.onUpdate(dx, dy, value(DismissDirection.endToStart));
         tower.dispatch(SetAllAnimationValues(IconPosition.LEFT, CardPosition.Left, dx, dy));
-        return;
+        animationController.animateTo(value(DismissDirection.endToStart), duration: Duration(milliseconds: 0));
       }
     }
+    /// Diagonal Heptic
+    if(!reverse && widget.controlHeptic){
+      if(oldDirection != i && ((dx.abs() >= HORIZONTALSWIPETHRESH && 
+      (i == 0 || i == 1)) || (dy.abs() > TOPSWIPETHRESH 
+      && i == 3) || (dy.abs() > BOTTOMSWIPETHRESH && i == 2))){
+        Functions.hapticSwipeVibrate();
+      }
+    }
+
+    oldDirection = i;
+
+    return;
   }
 
   /// Values of the animation system get updated externally
-  Future<void> _onFill(double? fill, {IconPosition? newIconPosition, CardPosition? newCardPosition, Duration? duration}) async {
-    animationController.animateTo(fill ?? 0.5, duration: duration, curve: Curves.easeInOutCubic);
+  Future<void> _onFill(DismissDirection direction, double dx, double? fill, {Duration? duration}) async {
     tower.dispatch(
       SetAllAnimationValues(
-        newIconPosition != null ? newIconPosition : tower.state.iconPosition,
-        newCardPosition != null ? newCardPosition : tower.state.cardPosition, 0, 0,
+        iconPosition(direction),
+        cardPosition(direction, dx), 
+        tower.state.dx, 
+        tower.state.dy,
         nullFill: fill == null,
       )
     );
+    await Future.delayed(Duration(milliseconds: 20));
+    animationController.animateTo(fill ?? 0.5, duration: duration, curve: Curves.easeInOutCubic);
     await widget.delegate.onFill(fill, tower.state);
     return;
   }
 
-  Future<bool> _onComplete(OverlayDelegate? overlay, {Future<void> Function()? reverse, List<dynamic>? args}) async {
+  Future<bool> _onComplete(DismissDirection direction, double dx, {Duration? duration, OverlayDelegate? overlay, Future<void> Function()? reverse, List<dynamic>? args}) async {
     Completer<bool> completer = Completer();
-    // Research into when this function runs
+
+    tower.dispatch(
+      SetAllAnimationValues(
+        iconPosition(direction),
+        cardPosition(direction, dx), 
+        tower.state.dx, 
+        tower.state.dy,
+        nullFill: false
+      )
+    );
+    await Future.delayed(Duration(milliseconds: 20));
+
     animationController.forward(from: animationController.value).whenComplete(() async {
-        bool result = await widget.delegate.onComplete(overlay, reverse: reverse, args: args);
-        completer.complete(result);
+      bool result = await widget.delegate.onComplete(tower.state, overlay: overlay, reverse: reverse, args: args ?? []);
+      completer.complete(result);
     });
     return completer.future;
   }
 
   @override
   Widget build(BuildContext context) {
-    return StoreProvider(
-      store: tower,
-      child: StoreConnector<AnimationSystemState, AnimationSystemState>(
-        converter: (store) => store.state,
-        builder: (context, state) {
-          return AnimatedBuilder(
-            animation: animationController,
-            builder: (context, _) {
-              return widget.delegate.build(context, state, animationController.value);
-            }
-          );
-        }
-      )
+    return AnimatedBuilder(
+      animation: animationController,
+      builder: (context, _) {
+        return widget.delegate.build(context, tower.state, animationController.value);
+      }
     );
   }
 }
 
+
 class AnimationSystemController extends ChangeNotifier{
-  late _AnimationSystemDelegateBuilderState? _state;
+  _AnimationSystemDelegateBuilderState? _state;
 
   //Bind to state
   void _bind(_AnimationSystemDelegateBuilderState bind) => _state = bind;
@@ -222,14 +324,18 @@ class AnimationSystemController extends ChangeNotifier{
   //Called to notify all listners
   void _update() => notifyListeners();
 
+  bool isBinded() => _state != null;
+
   //Update on update of dx and dy values inside of the animation state
-  void onUpdate(double dx, double dy, Function(DismissDirection) value) => _state != null ? _state!._onUpdate(dx, dy, value) : null;
+  void onUpdate(double dx, double dy, Function(DismissDirection) value, [bool reverse = false]) => _state != null ? _state!._onUpdate(dx, dy, value, reverse) : null;
+
+  Future<void> reverse() async => _state != null ? await _state!._reverse() : null;
 
   //Can be called when onSwipe is called to update the animation state
-  Future<void>? onFill(double? fill, {IconPosition? newIconPosition, CardPosition? newCardPosition, Duration? duration}) async => _state != null ? await _state!._onFill(fill, newIconPosition: newIconPosition, newCardPosition: newCardPosition, duration: duration) : null;
+  Future<void>? onFill(DismissDirection direction, double dx, double? fill, {Duration? duration}) async => _state != null ? await _state!._onFill(direction, dx, fill, duration: duration) : null;
 
   // Completes the current animation
-  Future<bool>? onComplete({OverlayDelegate? overlay, Future<void> Function()? reverse, List<dynamic>? args}) => _state != null ? _state!._onComplete(overlay, reverse: reverse, args: args) : null;
+  Future<bool>? onComplete(DismissDirection direction, double dx, {OverlayDelegate? overlay, Future<void> Function()? reverse, List<dynamic>? args}) => _state != null ? _state!._onComplete(direction, dx, overlay: overlay, reverse: reverse, args: args) : null;
 
   void reset() => _state != null ? _state!.reset() : null;